+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ CHANGELOG
 ==========
 
 * enhancement: Workflow: Specify tasks from which training/tuning operator to transform/deploy in related operators
+* feature: Supporting inter-container traffic encryption flag
 
 1.17.0
 ======